+++ src/components/Organisms/MarketingPreferencesDS/MarketingPreferencesDS.md
@@ -1,0 +1,8 @@
+Marketing Preferences DS
+
+```js
+import MarketingPreferencesDSForm from './MarketingPreferencesDSForm';
+
+<MarketingPreferencesDSForm />
+
+```
