+++ src/components/Organisms/MarketingPreferencesDS/assets/checkbox-white-tick.png
