--- src/components/Organisms/MarketingPreferencesRHF/MarketingPreferencesRHF.md
@@ -1,8 +1,0 @@
-Marketing Preferences RHF
-
-```js
-import MarketingPreferencesRHFForm from './MarketingPreferencesRHFForm';
-
-<MarketingPreferencesRHFForm />
-
-```
